# laravel_project/laravel/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel OpenApi Demo Documentation",
        "contact": {
            "email": "admin@admin.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8080/api",
            "description": "NGNIX serwer"
        }
    ],
    "paths": {
        "/user/subscribed_currencies": {
            "get": {
                "tags": [
                    "SubscribedCurrency"
                ],
                "summary": "check account balance",
                "operationId": "b730e34daa254958ff4c4efaaa0166d5",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "SubscribedCurrency"
                ],
                "summary": "Add Subscribed Currency",
                "operationId": "726f626d49dab0b51feb64e9fd766211",
                "requestBody": {
                    "description": "pass data to create Subscribed Currency record",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "subscribed_currency_name": {
                                        "description": "currency name",
                                        "type": "string",
                                        "example": "euro"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Unprocessable entity"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "SubscribedCurrency"
                ],
                "summary": "deletes all subscribed currency records of particular user",
                "operationId": "6fe3ac175e7abb65537b2eb3a67f9770",
                "responses": {
                    "204": {
                        "description": "No Content",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/user/subscribed_currencies/{currency_name}": {
            "delete": {
                "tags": [
                    "SubscribedCurrency"
                ],
                "summary": "deletes one subscribed currency record of particular user",
                "operationId": "ef22d6748154591c151a8ec585588ab7",
                "parameters": [
                    {
                        "name": "currency_name",
                        "in": "path",
                        "description": "currency_name for example: euro",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/user/register": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Register new user",
                "operationId": "414490d385bfef850e942b702c427190",
                "requestBody": {
                    "description": "pass data to create user record",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "user name",
                                        "type": "string",
                                        "example": "Jan"
                                    },
                                    "email": {
                                        "description": "user email",
                                        "type": "string",
                                        "example": "jankowalski@wp.pl"
                                    },
                                    "password": {
                                        "description": "user password",
                                        "type": "string",
                                        "example": "strongpassword123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Unprocessable entity"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/user/login": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Login user",
                "operationId": "8d892ea7d11a8c8ce720ff1ef49c3f80",
                "requestBody": {
                    "description": "pass data to Login user and get JWT token",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "description": "user email",
                                        "type": "string",
                                        "example": "jankowalski@wp.pl"
                                    },
                                    "password": {
                                        "description": "user password",
                                        "type": "string",
                                        "example": "strongpassword123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Unprocessable entity"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/user": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "view user",
                "operationId": "bdba3575e04a03e9b334f1813ad7cb75",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "security": [
        {
            "sanctum": []
        }
    ]
}